modules/location-grpc-kafka/requirements.txt
grpcio == 1.34.0
grpcio-tools == 1.34.0
kafka-python == 2.0.2
SQLAlchemy==1.3.19